[I25-178] refactor : /components/project/page.tsx 가 항상 build되는 문제 해결
